# src/back/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.CreateLabRequest:
    properties:
      course_id:
        type: integer
      deadline:
        type: string
      description:
        type: string
      github_url:
        type: string
      max_score:
        type: integer
      number:
        type: integer
      title:
        type: string
    required:
    - course_id
    - description
    - max_score
    - number
    - title
    type: object
  handlers.CreateLectureRequest:
    properties:
      content:
        type: string
      course_id:
        type: integer
      github_url:
        type: string
      title:
        type: string
      week:
        type: integer
    required:
    - content
    - course_id
    - title
    - week
    type: object
...
  handlers.UpdateLabRequest:
    properties:
      course_id:
        type: integer
      deadline:
        type: string
      description:
        type: string
      github_url:
        type: string
      max_score:
        type: integer
      number:
        type: integer
      title:
        type: string
    required:
    - course_id
    - description
    - max_score
    - number
    - title
    type: object
  handlers.UpdateLectureRequest:
    properties:
      content:
        type: string
      course_id:
        type: integer
      github_url:
        type: string
      title:
        type: string
      week:
        type: integer
    required:
    - content
    - course_id
    - title
    - week
    type: object
...
  models.Lab:
    properties:
      course_id:
        type: integer
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      github_url:
        type: string
      id:
        type: integer
      max_score:
        type: integer
      number:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.Lecture:
    properties:
      content:
        type: string
      course_id:
        type: integer
      created_at:
        type: string
      github_url:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      week:
        type: integer
    type: object
host: localhost:8443
info:
  contact:
    email: support@laritmo.local
    name: API Support
  description: Educational portal API with course management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Laritmo API
  version: "1.0"
paths:
...
  /api/admin/labs:
    post:
      consumes:
      - application/json
      description: Create a new lab
      parameters:
      - description: Lab data
        in: body
        name: lab
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateLabRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Lab'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create lab
      tags:
      - admin-labs
  /api/admin/labs/{id}:
    delete:
      description: Delete lab by ID
      parameters:
      - description: Lab ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete lab
      tags:
      - admin-labs
    put:
      consumes:
      - application/json
      description: Update lab by ID
      parameters:
      - description: Lab ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lab data
        in: body
        name: lab
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLabRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update lab
      tags:
      - admin-labs
  /api/admin/lectures:
    post:
      consumes:
      - application/json
      description: Create a new lecture
      parameters:
      - description: Lecture data
        in: body
        name: lecture
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateLectureRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Lecture'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create lecture
      tags:
      - admin-lectures
  /api/admin/lectures/{id}:
    delete:
      description: Delete lecture by ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete lecture
      tags:
      - admin-lectures
    put:
      consumes:
      - application/json
      description: Update lecture by ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      - description: Lecture data
        in: body
        name: lecture
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateLectureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update lecture
      tags:
      - admin-lectures
...
  /api/labs:
    get:
      description: Get list of all labs with optional course filter
      parameters:
      - description: Course ID filter
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Lab'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all labs
      tags:
      - labs
  /api/labs/{id}:
    get:
      description: Get lab details by ID
      parameters:
      - description: Lab ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lab'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get lab by ID
      tags:
      - labs
  /api/lectures:
    get:
      description: Get list of all lectures with optional course filter
      parameters:
      - description: Course ID filter
        in: query
        name: course_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Lecture'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all lectures
      tags:
      - lectures
  /api/lectures/{id}:
    get:
      description: Get lecture details by ID
      parameters:
      - description: Lecture ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Lecture'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get lecture by ID
      tags:
      - lectures
...
schemes:
- https
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"